1.完成调试代码，最佳准确率是 36%

+++ mnist/shared.cu
@@ -12,7 +12,7 @@ __device__ float fc2[C][H];
 __device__ float b2[C];
 __device__ float d_loss;
 
-__device__ int d_count;
+__device__ float d_count;
 
 
 float h_fc1[H][D];
@@ -20,7 +20,7 @@ float h_b1[H];
 float h_fc2[C][H];
 float h_b2[C];
 float h_loss;
-int h_count;
+float h_count;
 
 __device__ float d_g_d_fc1_w[H][D]; // device(d) global(g) deviration(d)
 __device__ float d_g_d_fc1_b[H];
